while showing ded

+++ derivative/differentiator.h
@@ -8,4 +8,4 @@ bool optimizeExpression(tree_t* derivative, treeNode_t* subTreeRoot, bool needDu
 tree_t tangentExpression(tree_t* expression, tree_t* derivative, tree_t* tailorX0, const char* variableToDiff);
 
 tree_t tailorExpansion(tree_t* expression, const char* variableToDiff, tree_t* tailorX0, tree_t* tailorOrder);
-int calculateSubTree (treeNode_t* subTreeRoot);
+double calculateSubTree (treeNode_t* subTreeRoot);
