+++ derivative/expressionTree.h
@@ -2,7 +2,7 @@
 
 treeNode_t* readExpression(tree_t* expression, char* buffer, size_t* curBufferPos);
 
-treeNode_t* createNewNodeNumber  (int value,  treeNode_t* left, treeNode_t* right);
+treeNode_t* createNewNodeNumber  (double value,  treeNode_t* left, treeNode_t* right);
 treeNode_t* createNewNodeVariable(const char* type, treeNode_t* left, treeNode_t* right);
 treeNode_t* createNewNodeOperator(const char* type, treeNode_t* left, treeNode_t* right);
