+++ general/mathHandlers.h
@@ -3,24 +3,24 @@
 
 #include "generalDifferentiator.h" 
 
-int addH(int* params);
-int subH(int* params);
-int mulH(int* params);
-int divH(int* params);
-int powH(int* params);
-int sinH(int* params);
-int cosH(int* params);
-int tanH(int* params);
-int cotH(int* params);
-int lnH (int* params);
-int shH (int* params);
-int chH (int* params);
-int arcsinH (int* params);
-int arccosH (int* params);
-int arctanH (int* params);
+double addH(double* params);
+double subH(double* params);
+double mulH(double* params);
+double divH(double* params);
+double powH(double* params);
+double sinH(double* params);
+double cosH(double* params);
+double tanH(double* params);
+double cotH(double* params);
+double lnH (double* params);
+double shH (double* params);
+double chH (double* params);
+double arcsinH (double* params);
+double arccosH (double* params);
+double arctanH (double* params);
 
-int sqrtH (int* params);
+double sqrtH (double* params);
 
-int factorial(int n);
+double factorial(double n);
 
 #endif /* MATH_HANDLERS_H */
